update where root values are assigned from

+++ unicorn/include/def.hpp
@@ -73,8 +73,8 @@ extern std::string HIST_FILE;
  * - "new_dst": whether the destination node is new. never-before-seen.
  */
 struct RootPair {
-	uint32_t root;
-	uint32_t order;
+	unsigned long root;
+	unsigned long order;
 };
 
 typedef struct edge_label {
